requirements.txt
python-dotenv
langchain
langchain-community
langchain-core
openai
chromadb
textwrap3